requirements.txt
streamlit
bcrypt
sqlalchemy
psycopg2-binary
pandas
plotly
folium>=0.14.0
streamlit-folium>=0.15.0
geopy>=2.3.0
